// client/state/User/user.action-types.ts
export enum ActionTypes {
  USER_LOGIN_START = 'USER_LOGIN_START',
  USER_LOGIN_SUCCESS = 'USER_LOGIN_SUCCESS',
  USER_LOGIN_ERROR = 'USER_LOGIN_ERROR',

  USER_REGISTER_START = 'USER_REGISTER_START',
  USER_REGISTER_SUCCESS = 'USER_REGISTER_SUCCESS',
  USER_REGISTER_ERROR = 'USER_REGISTER_ERROR',

  USER_UPDATE_START = 'USER_UPDATE_START',
  USER_UPDATE_SUCCESS = 'USER_UPDATE_SUCCESS',
  USER_UPDATE_ERROR = 'USER_REGISTER_ERROR',

  GET_CURRENT_USER = 'GET_CURRENT_USER',
  USER_LOGOUT = 'USER_LOGOUT',
  CLEAN_USER_ERRORS = 'CLEAN_USER_ERRORS',
}